Allow general hash table size up to 1 TB
This patch keeps the size of the hash table general, but allows its size to be up to 1 TB instead of 128 GB. This is achieved by adjusting the multiplication operation that scales the hash key to a cluster index, making use of the fact that the size of the hash table is specified in megabytes (i.e. a multiple of 2^15 clusters).

The implementation is kept general to allow future changes of the size of a cluster. The maximum hash table size can also be easily changed by modifying the SizeLog constant in tt.h.

For a general implementation, it was easier to use the higher bits of the hash key to determine the cluster index and to use the lower 16 bits to distinguish keys within a cluster.

In search.c the excludedMove is xor'ed into the highest 16 bits of the hashkey to avoid that the positions with and without excluded move are mapped to the same cluster.

Since it is sufficient that the transposition table is aligned on clusters, there is no need to consider the (typical) size of a cache line. Instead, we make sure that the size of a cluster is a power of two.

Bench: 5291480